# services/matematika/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для генерации финансовых выписок и транзакций. Сервис позволяет генерировать реалистичные финансовые выписки на основе заданных параметров (оборот, желаемая прибыль, модель бизнеса).",
        "title": "Matematika API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/IbadT/business_bank_back",
            "email": "support@matematika.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Генерирует финансовую выписку с транзакциями на основе переданных параметров. Поддерживает модели B2C и B2B, позволяет задавать желаемый процент прибыли, начальный баланс и дополнительные кастомные данные.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statements"
                ],
                "summary": "Генерация финансовой выписки",
                "parameters": [
                    {
                        "description": "Параметры генерации выписки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная генерация выписки",
                        "schema": {
                            "$ref": "#/definitions/dto.GenerateResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос - ошибки валидации входных параметров",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации - транзакция приведет к отрицательному балансу",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/holidays": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает список всех праздников для указанного года. Формат года: YYYY (например, 2024). Требуется авторизация.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holidays"
                ],
                "summary": "Получить список всех праздников для указанного года",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Год для получения праздников в формате YYYY",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение списка праздников",
                        "schema": {
                            "$ref": "#/definitions/dto.GetHolidaysResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос - ошибки валидации входных параметров",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет новый праздник в базу данных. Требуется авторизация. Поддерживаемые страны: RU, US.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holidays"
                ],
                "summary": "Добавление праздника",
                "parameters": [
                    {
                        "description": "Данные для добавления праздника",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.HolidayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное добавление праздника",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос - ошибки валидации входных параметров",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/holidays/is-holiday": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Проверяет, является ли указанная дата праздником. Формат даты: YYYY-MM-DD (например, 2024-12-15). Требуется авторизация.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holidays"
                ],
                "summary": "Проверка является ли дата праздником",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата для проверки в формате YYYY-MM-DD",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат проверки даты",
                        "schema": {
                            "$ref": "#/definitions/dto.IsHolidayResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос - ошибки валидации входных параметров",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/holidays/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет существующий праздник в базе данных по его ID. Требуется авторизация.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holidays"
                ],
                "summary": "Обновление праздника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID праздника",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления праздника",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.HolidayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление праздника",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос - ошибки валидации входных параметров",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Праздник не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет праздник из базы данных по его ID. Требуется авторизация.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holidays"
                ],
                "summary": "Удаление праздника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID праздника",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление праздника",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос - ошибки валидации входных параметров",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Праздник не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.GenerateRequest": {
            "description": "Параметры для генерации финансовой выписки",
            "type": "object",
            "required": [
                "desiredProfitPercent",
                "initialBalance",
                "model",
                "month",
                "turnover",
                "year"
            ],
            "properties": {
                "customData": {
                    "$ref": "#/definitions/dto.CustomData"
                },
                "desiredProfitPercent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 10
                },
                "initialBalance": {
                    "type": "number",
                    "minimum": 0,
                    "example": 300000
                },
                "model": {
                    "type": "string",
                    "enum": [
                        "B2C",
                        "B2B"
                    ],
                    "example": "B2C"
                },
                "month": {
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1,
                    "example": 12
                },
                "scaleFactor": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 1,
                    "example": 1
                },
                "turnover": {
                    "type": "number",
                    "minimum": 0,
                    "example": 100000
                },
                "year": {
                    "type": "integer",
                    "maximum": 2100,
                    "minimum": 2000,
                    "example": 2024
                }
            }
        },
...
        "dto.MessageResponse": {
            "description": "Стандартный ответ с сообщением об успешной операции",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите только токен (без Bearer). Получите токен через /api/login или /api/register. Swagger автоматически добавит префикс \"Bearer \"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}